Updated portrait and added print-out screen shots
The 6sp and Xsmax screen shots are new; the former because the app now
runs in native resolution, and the latter thanks to a new skin for the
X series phones that was donated. I also replaced the 5s screen shot,
although I don't think the skin actually changed on that one.

Added print-out screen shots for all four sizes. The 4s print-out is
fake, created by cutting off the bottom of the 5s screen shot, because
I have nothing to test 4s screens any more; the other three print-out
screen shots are real, taken on a real 5s and on 8p and 11ProMax
simulators.