Added SkinMuncher input files
For generating landscape and shortened skins for the Android and iOS
versions. The skin munching logic is currently in the stand-alone
SkinMuncher program, but should be built into the Plus42 executable, so
only the standard Plus42 skin needs to be built into the executable, and
the landscape and shortened versions can be generated on the fly.